Move images into ggrc-client directory
