+++ src/ggrc-client/images/ajax-loader.gif
