+++ src/ggrc-client/images/bgnds/drag-h.png
