+++ src/ggrc-client/images/bgnds/drag-v.png
