+++ src/ggrc-client/images/favicon.ico
