+++ src/ggrc-client/images/magnifying-glass.png
