+++ src/ggrc/assets/javascripts/entrypoints/styles.js
@@ -3,5 +3,5 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import '../../images/favicon.ico';
+import '../../../../ggrc-client/images/favicon.ico';
 import '../../stylesheets/dashboard.scss';
