+++ src/ggrc/assets/stylesheets/_common.scss
@@ -58,7 +58,7 @@ a {
 .objective-selector {
   .search-icon {
     @include background-size(21px 16px);
-    background-image: url("../images/magnifying-glass.png");
+    background-image: url("../../../ggrc-client/images/magnifying-glass.png");
     background-repeat: no-repeat;
     background-position: right center;
     padding-right: 28px;
@@ -173,14 +173,14 @@ a {
   &.pending-ajax {
     box-shadow: none;
     cursor: default;
-    background:$widgetHeaderBgnd url('../images/ajax-loader.gif') no-repeat 5px 6px;
+    background:$widgetHeaderBgnd url('../../../ggrc-client/images/ajax-loader.gif') no-repeat 5px 6px;
     border: 1px solid $ajaxBtnDisabled;
     text-shadow: none;
     padding-left: 28px;
 
     &:active {
       box-shadow: none;
-      background:$widgetHeaderBgnd url('../images/ajax-loader.gif') no-repeat 5px 5px;
+      background:$widgetHeaderBgnd url('../../../ggrc-client/images/ajax-loader.gif') no-repeat 5px 5px;
       border: 1px solid $ajaxBtnDisabled;
     }
   }
@@ -280,7 +280,7 @@ select.input-block-level {
     background-color:#fff;
     border-top:1px solid #aaa;
     border-bottom:1px solid #aaa;
-    background-image: url("../images/bgnds/drag-h.png");
+    background-image: url("../../../ggrc-client/images/bgnds/drag-h.png");
     background-repeat:no-repeat;
     background-position:50% 1px;
     &:hover {
@@ -347,7 +347,7 @@ select.input-block-level {
 
 // AJAX loading gif
 input.disabled.pending-ajax{
-  background: $lightGray url("../images/ajax-loader.gif") no-repeat right center;
+  background: $lightGray url("../../../ggrc-client/images/ajax-loader.gif") no-repeat right center;
 }
 
 // Left dropdown positioning
